services/map-service/requirements.txt
flask
numpy
pandas
openpyxl
python-dotenv